requirements.txt
torch
torchvision
numpy
pandas
tqdm

opencv-python-headless
Pillow
lmdb

scipy
scikit-image
shapely
svgpathtools

matplotlib
tensorboardX

flask
